+++ test/scjsv/core_test.clj
@@ -38,7 +38,7 @@
     (validate-with-explicit-factory valid) => nil
     (validate-with-explicit-factory invalid) =not=> nil
 
-    (fact "validation errors are lovey clojure maps"
+    (fact "validation errors are lovely clojure maps"
       (validate invalid)
       => [{:domain "validation"
            :instance {:pointer "/shipping_address"}
@@ -49,3 +49,11 @@
            :required ["city" "state" "street_address"]
            :schema {:loadingURI "#"
                     :pointer "/definitions/address"}}])))
+
+(fact "Validating a schema that refers to itself"
+  (let [schema (slurp (io/resource "scjsv/with_id.json"))
+        default-validate (v/validator schema)
+        inline-validate (v/validator schema {:dereferencing :inline})
+        valid {:foo "foo"}]
+    (default-validate valid) =not=> nil
+    (inline-validate valid) => nil))